feat: rename from balance -> nft-balance
