+++ packages/indexer/src/jobs/websocket-events/index.ts
@@ -1,4 +1,4 @@
 import "@/jobs/websocket-events/ask-websocket-events-trigger-queue";
 import "@/jobs/websocket-events/bid-websocket-events-trigger-queue";
-import "@/jobs/websocket-events/balance-event-websocket-events-trigger-queue";
+import "@/jobs/websocket-events/nft-balance-event-websocket-events-trigger-queue";
 import "@/jobs/websocket-events/new-top-bid-trigger-queue";
